+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # RELEASE HISTORY
 
+### 5.7.14 / May 6, 2020
+* fixed issue when user property is object of custom class
+* fixed issue with class casting when Bunlde Name and Bundle Executable Name have different values
+
 ### 5.7.13 / April 30, 2020
 * fixed the FileService `remove` method response
 * tests fixed